Add progress bar to autoinstall script

+++ lib/python/Screens/StartWizard.py
@@ -10,6 +10,7 @@ except:
 	OverscanWizard = None
 
 from Components.Pixmap import Pixmap
+from Components.ProgressBar import ProgressBar
 from Components.ScrollLabel import ScrollLabel
 from Components.config import config, ConfigBoolean, configfile
 from LanguageSelection import LanguageWizard
@@ -62,20 +63,39 @@ class AutoInstallWizard(Screen):
 		<panel position="right" size="5%,*"/>
 		<panel position="top" size="*,5%"/>
 		<panel position="bottom" size="*,5%"/>
+		<widget name="progress" position="top" size="*,24" backgroundColor="#00242424"/>
 		<widget name="AboutScrollLabel" font="Fixed;20" position="fill"/>
 	</screen>"""
 	def __init__(self, session):
 		Screen.__init__(self, session)
+		self["progress"] = ProgressBar()
+		self["progress"].setRange((0, 100))
+		self["progress"].setValue(0)
 		self["AboutScrollLabel"] = ScrollLabel(_("Please wait"), showscrollbar=False)
+
+		self.logfile = open('/home/root/autoinstall.log', 'w')
 		self.container = eConsoleAppContainer()
 		self.container.appClosed.append(self.appClosed)
 		self.container.dataAvail.append(self.dataAvail)
-		self.onLayoutFinish.append(self.run_console)
+		self.counter = 0
+
+		macaddr = open('/sys/class/net/eth0/address', 'r').readline().strip().replace(":", "")
+		backupdir = os.path.join(os.sep, config.plugins.autobackup.where.value or '/media/hdd', 'backup')
+		autoinstallfiles = [x for x in os.path.join(os.sep, backupdir, '%s%s' % ('autoinstall', macaddr)), os.path.join(os.sep, backupdir, 'autoinstall') if os.path.isfile(x)]
+		if autoinstallfiles:
+			self.packages = [x.strip() for x in open(autoinstallfiles[0]).readlines()]
+			self.totalpackages = len(self.packages)
+			self.onLayoutFinish.append(self.run_console)
+		else:
+			self.close()
 
 	def run_console(self):
-		self["AboutScrollLabel"].setText("")
+		if not self.counter:
+			self["AboutScrollLabel"].setText("")
+		self.counter += 1
+		self["progress"].setValue(100 * self.counter/self.totalpackages)
 		try:
-			if self.container.execute("/etc/init.d/autoinstall.sh"):
+			if self.container.execute("/etc/init.d/autoinstall.sh %s" % self.packages.pop()):
 				raise Exception, "failed to execute autoinstall.sh script"
 				self.appClosed(True)
 		except Exception, e:
@@ -83,15 +103,20 @@ class AutoInstallWizard(Screen):
 
 	def dataAvail(self, data):
 		self["AboutScrollLabel"].appendText(data)
+		self.logfile.write(data)
 
 	def appClosed(self, retval):
 		if retval:
 			self["AboutScrollLabel"].setText(_("An error occurred - Please try again later"))
-		self.container.appClosed.remove(self.appClosed)
-		self.container.dataAvail.remove(self.dataAvail)
-		self.container = None
-		os.remove("/etc/.doAutoinstall")
-		self.close(3)
+		if self.packages:
+			self.run_console()
+		else:
+			self.container.appClosed.remove(self.appClosed)
+			self.container.dataAvail.remove(self.dataAvail)
+			self.container = None
+			self.logfile.close()
+			os.remove("/etc/.doAutoinstall")
+			self.close(3)
 
 if not os.path.isfile("/etc/installed"):
 	from Components.Console import Console
